workspace/pipeline-output-servers/qwen-plus/mcp_mysql_database_manager/refined/requirements.txt
mcp[cli]
mysql-connector-python
typing_extensions